Don't test unnecessary things
Since we're now merging the default options and user options, we don't
really need to check the conditional assignment. That would just be
testing that the `...` operator  works.

Additionally, we don't really need to test that the alias is being added
to the poolCredentials property, and there's no point in having
pointless tests.